Remove figlet problem short-circuit
If there’s a problem rendering the ASCII-art banner, the program will exit
before doing anything useful and with no error message.

Arguably, a user would prefer the program to continue setting up the ReactXP
project even if they don’t get to see the banner in its full glory.

This change allows the program to continue executing even in the unlikely event
that figlet reports a problem, instead falling-back to a regular text output
for the banner.